Ensure we are using the log4j implementation that we think we are
SLF4J and the parent/child pom relationship was not being
used correctly, and as a result, was falling back to a log4j1
version and not the log4j2 version as stated in the parent pom.
This was caused by the following commit 2b99b895f .

Many thanks to Dr Adam Thorn <alt36@cam.ac.uk> for spotting and
investigating this.